organizes file structure
